// src/components/nav-menu/nav-menu-enum.ts
export enum optionsEnum {
  about = 'about',
  myProjects = 'myProjects',
  nextLevelWeek = 'nextLevelWeek',
  experience = 'experience',
  contact = 'contact'
}

export enum optionsEnumLabel {
  about = 'About',
  myProjects = 'My Projects',
  nextLevelWeek = 'Next Level Week',
  experience = 'Work Experience',
  contact = 'Contact'
}